+++ src/core/dev/dpcp/cq_mgr_rx_inl.h
@@ -34,8 +34,9 @@
 #ifndef CQ_MGR_INL_H
 #define CQ_MGR_INL_H
 
+#ifdef DEFINED_DPCP_PATH_RX
 #include "cq_mgr_rx.h"
-#include "ring_simple.h"
+#include "dev/ring_simple.h"
 #include "util/utils.h"
 #include <netinet/ip.h>
 #include <netinet/ip6.h>
@@ -110,4 +111,5 @@ inline bool is_eth_tcp_frame(mem_buf_desc_t *buff)
     return false;
 }
 
+#endif // DEFINED_DPCP_PATH_RX
 #endif // CQ_MGR_INL_H
