+++ src/core/dev/dpcp/cq_mgr_rx_strq.cpp
@@ -31,14 +31,14 @@
  * SOFTWARE.
  */
 
+#include "config.h"
+#ifdef DEFINED_DPCP_PATH_RX
+
 #include "cq_mgr_rx_strq.h"
-
-#if defined(DEFINED_DIRECT_VERBS)
-
 #include <util/valgrind.h>
 #include "cq_mgr_rx_inl.h"
-#include "hw_queue_rx.h"
-#include "ring_simple.h"
+#include "dev/dpcp/hw_queue_rx_dpcp.h"
+#include "dev/ring_simple.h"
 #include <cinttypes>
 
 #define MODULE_NAME "cq_mgr_rx_strq"
@@ -526,4 +526,4 @@ void cq_mgr_rx_strq::reclaim_recv_buffer_helper(mem_buf_desc_t *buff)
     }
 }
 
-#endif /* DEFINED_DIRECT_VERBS */
+#endif // DEFINED_DPCP_PATH_RX
