+++ src/core/dev/ring_bond.h
@@ -112,19 +112,20 @@ public:
     bool tls_tx_supported(void) override;
     bool tls_rx_supported(void) override;
     xlio_tis *tls_context_setup_tx(const xlio_tls_info *info) override;
+    void tls_context_resync_tx(const xlio_tls_info *info, xlio_tis *tis, bool skip_static) override;
+    void tls_release_tis(xlio_tis *tis) override;
+    void tls_tx_post_dump_wqe(xlio_tis *tis, void *addr, uint32_t len, uint32_t lkey,
+                              bool first) override;
+#ifdef DEFINED_DPCP_PATH_RX
     xlio_tir *tls_create_tir(bool cached) override;
     int tls_context_setup_rx(xlio_tir *tir, const xlio_tls_info *info, uint32_t next_record_tcp_sn,
                              xlio_comp_cb_t callback, void *callback_arg) override;
     rfs_rule *tls_rx_create_rule(const flow_tuple &flow_spec_5t, xlio_tir *tir) override;
-    void tls_context_resync_tx(const xlio_tls_info *info, xlio_tis *tis, bool skip_static) override;
     void tls_resync_rx(xlio_tir *tir, const xlio_tls_info *info,
                        uint32_t hw_resync_tcp_sn) override;
     void tls_get_progress_params_rx(xlio_tir *tir, void *buf, uint32_t lkey) override;
-    void tls_release_tis(xlio_tis *tis) override;
     void tls_release_tir(xlio_tir *tir) override;
-    void tls_tx_post_dump_wqe(xlio_tis *tis, void *addr, uint32_t len, uint32_t lkey,
-                              bool first) override;
-
+#endif // DEFINED_DPCP_PATH_RX
 #endif /* DEFINED_UTLS */
 
     void slave_create(int if_index);
