+++ src/core/util/sys_vars.h
@@ -452,7 +452,6 @@ public:
     char internal_thread_affinity_str[FILENAME_MAX];
     cpu_set_t internal_thread_affinity;
     bool internal_thread_arm_cq_enabled;
-    bool doca_rx;
     bool doca_tx;
     skip_poll_in_rx_t skip_poll_in_rx;
     multilock_t multilock;
